fix(clients): abort signed-url fetch on local disk failures
