+++ src/shell/platform/fetchSignedUrl.test.ts
@@ -161,88 +161,4 @@ describe("fetchSignedUrl", () => {
     expect(result).toEqual({ ok: true, data: undefined });
     expect(readFileSync(dest, "utf8")).toBe("abcdef");
   });
-
-  // Streaming to disk keeps peak memory at one chunk, not 2× the file size —
-  // large assets must not be buffered whole before writing.
-  it("streams each chunk to disk as it arrives instead of buffering the body", async () => {
-    const fs = makeMemoryFs();
-    await fs.mkdir("/assets", { recursive: true });
-    const chunkWrites: number[] = [];
-    const openWriteHandle = fs.openWriteHandle.bind(fs);
-    fs.openWriteHandle = async (path) => {
-      const handle = await openWriteHandle(path);
-      return {
-        write: async (chunk) => {
-          chunkWrites.push(chunk.length);
-          await handle.write(chunk);
-        },
-        close: handle.close,
-      };
-    };
-
-    const result = await fetchSignedUrl(
-      { dest: "/assets/file.bin", url },
-      {
-        fetch: makeDrippingBodyFetch(["ab", "cd", "ef"], 10),
-        fs,
-        stallTimeoutMs: 200,
-      },
-    );
-
-    expect(result).toEqual({ ok: true, data: undefined });
-    expect(chunkWrites).toEqual([2, 2, 2]);
-    expect(await fs.readFile("/assets/file.bin")).toBe("abcdef");
-    expect(await fs.pathExists("/assets/file.bin.part")).toBe(false);
-  });
-
-  // The stall clock measures the network, not the disk: a write that outlasts
-  // the window must not abort a download whose bytes keep arriving.
-  it("does not count slow disk writes toward the stall window", async () => {
-    const fs = makeMemoryFs();
-    await fs.mkdir("/assets", { recursive: true });
-    const openWriteHandle = fs.openWriteHandle.bind(fs);
-    fs.openWriteHandle = async (path) => {
-      const handle = await openWriteHandle(path);
-      return {
-        write: async (chunk) => {
-          await new Promise((resolve) => setTimeout(resolve, 80));
-          await handle.write(chunk);
-        },
-        close: handle.close,
-      };
-    };
-
-    const result = await fetchSignedUrl(
-      { dest: "/assets/file.bin", url },
-      {
-        fetch: makeDrippingBodyFetch(["ab", "cd"], 10),
-        fs,
-        stallTimeoutMs: 50,
-      },
-    );
-
-    expect(result).toEqual({ ok: true, data: undefined });
-    expect(await fs.readFile("/assets/file.bin")).toBe("abcd");
-  });
-
-  // Failed local writes must not reclassify as network stalls, and a partial
-  // file must not survive the failure.
-  it("returns a network error and removes the partial file when a write fails", async () => {
-    const fs = makeMemoryFs();
-    await fs.mkdir("/assets", { recursive: true });
-    const cause = new Error("disk full");
-    fs.openWriteHandle = async () => ({
-      write: () => Promise.reject(cause),
-      close: async () => {},
-    });
-
-    const result = await fetchSignedUrl(
-      { dest: "/assets/file.txt", url },
-      { fetch: asFetch(createFetchMock(new Response("bytes"))), fs },
-    );
-
-    expect(result).toEqual({ ok: false, error: { cause, kind: "network" } });
-    expect(await fs.pathExists("/assets/file.txt")).toBe(false);
-    expect(await fs.pathExists("/assets/file.txt.part")).toBe(false);
-  });
 });
